new changes

+++ database/migrations/2022_01_05_131812_create_sub_goals_table.php
@@ -18,7 +18,7 @@ class CreateSubGoalsTable extends Migration
             $table->foreignId('task_id')->constrained();
             $table->string('title');
             $table->string('description');
-            $table->unsignedBigInteger('rating');
+            $table->float('rating');
             $table->timestamps();
         });
     }
